Replaced the Java main method contract check with a more generic mechanism. Specifically, a .jarsrc file may now contain one or more "provides <cmdLine>" lines. For example, to indicate that a given .jar file provides a main method in class mypackage.MyClass, write
    provides main_class mypackage.MyClass

VeriFast will run the command line "main_class mypackage.MyClass > myjarfile_provide0.java". The "main_class" program generates a test class that tests the main method. This generated file is included in the verification.

The old syntax

    main-class mypackage.MyClass

is now interpreted as a shorthand for

    provides main_class mypackage.MyClass

The generated file is deleted by verifast.exe after a successful verification, but not by vfide.exe.
TODO: Add a java_card_applet script that generates a class that tests a Java Card applet's install method.