New buildsystem is done!
It went a bit out of hand but it has reached its target goals
and a few extras: it is faster, crossplatform, easyer to modify,
a lot cleaner, supports bytecode builds, supports debug builds,
and checks for (most) missing dependencies with proper errors.

There are a few things that still could be improved (e.g. the
LD_LIBRARY_PATH setup could be cleaner) and writing Makefiles
turns out to be a bit error-prone because Make does not have
proper typechecking (everything is a string). But it should
be good enough for now.

This commit only
- adds support for saving config to prevent having to set the
  Z3 path every build
- autodetects NOZ3=1 on mac
- adds dependency to gtksourceview2 on Windows-vfide (untested)
- adds proper error if gtksourceview2 is missing (untested)
- proper usage help.