vfide builds again.
For some reason Unix is already included and apparently both ocamlbuild
and ocamlc are not smart enough to handle multiple includes of the very
same file, so we remove the include of unix for the vfide build.

Also minor change: "cp -f" instead of "cp" such that it works too if
vfide is running while compiling vfide.