Fixed an unsoundness: octal numbers were lexed as decimal ones.
In C and Java, a number containing only digits but starting with zero
(e.g. "0123") is octal, just like a number with leading "0x" is
hexadecimal (eg "0x123").

Credits: thanks to Gijs for pointing out Java uses the same convention
for octal numbers. We thus do not need a (potentially cumbersome) case
distinction.