The bool.h mess is now gone.
bool.h is not C99 standard. As a result, every example that
used booleans and wanted to be compilable with gcc, shipped
a bool.h. This is messy. Also, since including bool.h was
ignored by VeriFast, this is an unsoundness that is not
clearly marked as such, "bool.h" looks pretty innocent.

Now, we use stdbool.h which is C99 compliant. VeriFast programs
that use booleans and include stdbool.h just compile with gcc
out of the box.

To avoid hard to fix regressions, we still support the feature
of secret includes, but made it such that it is obvious what
you are doing: write
  #include "include_ignored_by_verifast.h"
instead of the innocent looking
  #include "bool.h"