io: Merge commit '4c2b2d7': asserts and invariant with mixing two progresses
This is work from a weak ago. I merge it now because
it is the branch that is the closest to how we plan to continue.

The example uses split/join (not nested) where a simple "mix"
predicate is used in the invariant to interleave two progresses.
By obtaining the final token and the invariant, one can assert
constraints about the fysical state holds.

In a more real-world example, we expect that one does not (or
not often) need to write the invariant using a kind of generic
"mix" predicate, but usually can write the invariant in a
fashion more similar to 01-producer-consumer-no-iostyle/ combined
with using the progresses.