Found an unsoundness: there is a way to call a nonghost_callers_only lemma in a normal lemma.
lemma ...
nonghost_callers_only // (2)
...
{
  ... // <-- context is nonghost_callers_only
  produce_lemma_function_pointer_chunk(...):... {
     ... // <-- context should not be nonghost_callers_only (1)
     call();
     ...
  }
  {
     ... // <-- context is nonghost_callers_only
  }
}

A call to a nonghost_callers_only-lemma at (1) is currently allowed if (2) is nonghost_callers_only.
This allows to wrap a nonghost_callers_only lemma call inside a lemma function pointer chunk,
which can then later be called by a non-nonghost_callers_only lemma.
I think the fix is to always disallow nonghost_callers_only lemma calls at (1).

I wanted to implement this fix but it seems that it'll take longer than expected
and I'm behind schedule so I canceled trying to fix this for now.

The .c file in this commit shows an exploit that proves false.