Fix the type assigned to `UINT_MAX` and `UINTPTR_MAX`
It should be `unsigned int`, not `int` (or `long long int`).
This matters when used in a comparison (in non-ghost code).

In particular, per the C standard, `-1 < 3000000000` evaluates to
`true`, whereas `-1 < (unsigned)3000000000` evaluates to `false`.
In the former, `3000000000` is of type `long long int`.
In the latter, `-1` is converted to `unsigned int`, yielding
`4294967295`.