Fix unsoundness of combining lemma ptrs with nonghost_callers_only
It was possible to call a nonghost_callers_only lemma in a context
that is not nonghost_callers_only by abusing lemma function
pointer chunks.

The ``attack'' consisted of creating a lemma function pointer chunk
in a nonghost_callers_only lemma. The (first) body of the chunk
would be considered nonghost_callers_only because of the
nonghost_callers_only-ness of the outer lemma/realfunction,
and would thus be able to wrap a call to a nonghost_callers_only lemma.

A lemma that is not nonghost_callers_only could then indirectly
call the nonghost_callers_only through the lemma function pointer
chunk.

An example that proved false using the unsoundness is in
  lemmafuncptr_call_nonghost_in_producelemfunptr.c

The fix in this commit disallows calling nonghost_callers_only
lemmas in the (first) body of a lemma function pointer chunk.

This is implemented by creating a new leminfo for verifying
the body of the lemma function pointer chunk (instead of reusing
the leminfo of the outer lemma/function which might be
nonghost_callers_only) and track the necessary info to be able
to populate its fields.