build: win: Check sha256sum instead of choking on StartCom TLS cert
The HTTPS download from forge.ocamlcore.org fails because the
TLS certificate is not accepted. This breaks the Windows build.

The reason is unknown. The certificate is from StartCom, which is
not a trusted CA anymore, but currently does not seem te be revoked.
For some reason forge.ocamlcore.org is rejected by wget, but other
downloads with also a cert from StartCom just work.

I replaced HTTPS with HTTP to keep the build working and, for security,
I added a check to compare the sha256. In case you are worried
about this: I compared the sha256 with a download I had from
May 2016 and it's still the same. The right thing to do would be
to check against a (PGP) signature, but forge.ocamlcore.org does
not seem to sign their releases.