Add support for Z3 version 4.5
Main motivation for supporting Z3: its SAT solver is far superior to Redux's.
The new example bitops.c (which is SAT-heavy) verifies in .3s with Z3v4.5 and
in 8s with Redux.

The new example illustrates a way to reason about operations on bounded
integers (in particular: bitwise operations) in a bitwise fashion.