Smtlib dump
This defines a fake prover that always answers "unknown" but performs
a dump in SMTlib format in a file named "dump.smt2" as a side effect.

The main goal for this SMTlib output is to benchmark SMT solvers on
the kind of problems typical for Verifast.

Since Verifast workflows requires interaction with the prover, this
fake prover is not usable in standalone but needs to be combined with
a real prover. This commit provides combination plugins for Redux and
Z3 version 4.5.