Adding prover features
Each prover using the SMTLib interface declares a set of features and
from each node in the SMTLib AST, we can compute the set of features
required for this node to make sense. Provers do not receive
assumptions that they cannot understand. For now however, pushes and
pops are sent to the prover even if it does not declare the
incrementality feature.