Add generic pointee pred integer_ #139 #140
For now, it is used only when no legacy
pointee pred is available for the type. This
is the case specifically for (u)int128_t, as
well as for (u)int32_t under -target IP16.

Support for the generic pointee predicate
is very partial for now. For example, arrays are
not yet supported.