Improve ternary operator performance
VeriFast now keeps Redux from eagerly creating termnodes for
symbol applications in the branches of an if-then-else expression.
This dramatically improves performance in some cases, for example
when verifying 'scanf' call sites.

This commit also adds support for the 'c' format specifier in
scanf formats, and adds a partial spec for function 'setbuf' to stdio.h.

The new treatment of if-then-else expressions is less complete; this
required tweaking one test case.