Relax check of E in 'malloc(E * sizeof(T))'
VeriFast now allows E to be of any integer type, and checks that
the value of 'E * sizeof(T)' is within the limits of size_t.

This commit also adds lemma 'pointers_limits' and macro 'SIZE_MAX'
and updates Redux to evaluate integer division terms to their
literal value when the operands' literal values are known at the
point where the term is created. This eliminates the need for some
calls of lemma 'div_rem'.