Redux: preserve nonlinear products as term nodes
This commit improves Redux's treatment of nonlinear arithmetic.
Redux now generates a term node for each multiplication term
given to it (unless either operand's value is known). Furthermore,
it produces the equality of this term node with its reduced
version, where multiplication is distributed over addition.

That is, the term '(x + 1) * y' now produces a term node for
'(x + 1) * y', as well as the equality

    (x + 1) * y == x * y + y

This eliminates the need for the workaround where the user had to
define a fixpoint

    fixpoint int mul(int x, int y) { return x * y; }

to force Redux to produce a term node for the multiplication.