Produce length of updated C array contents
The statement `xs[i] = v;` consumes and produces the array chunk
for `xs`, but it does not apply autorules when producing it.
Therefore, the autolemma that produces the relationship between
the size of the array and the length of elements list is not
called. This patch produces that relationship directly.