Eliminate type assumptions in the provers
The provers incorrectly assumed that VeriFast would not
generate assumptions of the form t1 == t2 where t1 and t2 are
of different "inductive types". Furthermore, Redux assumed that
VeriFast would not generate assumptions equating any two of the
following:
- a boxed boolean
- a boxed number
- a value of an "inductive type".

VeriFast may indeed generate such assumptions, e.g. when
fixpoint default_value is used and when type "any" is used.

This commit refines the prover API so that the client can specify,
for each constructor or fixpoint symbol, which "inductive type" this
symbol pertains to. We call these "inductive subtypes", because they
are (overlapping) subtypes of the single prover type "inductive".

Redux now keeps track, for each value node, of a set of constructor
terms in the equivalence class, at most one per inductive subtype.

Since Redux infers "t == true" from "t != false", boxing a boolean
is no longer a no-op.

The Z3 adaptor has also been fixed so that it does not assume that
constructors of distinct inductive subtypes are disjoint.

Fixes #68.
Fixes #110.