Support treating structs as values
This change adds support for treating local structs whose address is not
taken as values instead of treating them as objects.  In particular, it
is possible to pass a struct by value to functions and predicates, and
to return them by value.

This change also adds support for "x.f" syntax for accessing
constructor arguments of values of inductive types that only have one
constructor.

Both changes cause the automatic generation of getter/setter function
symbols to read/write fields of the struct/constructor and axioms to
support them.  These getter/setter functions are used in r-values and
l-values.

Co-authored-by: Bart Jacobs <bart.jacobs@cs.kuleuven.be>