merged NewObject and NewGenericObject into one by making type arguments optional.
None means the object is either raw or a regular object creation. If the constructor has type parameters present they are initialised as java.lang.Object.

Some(targs) means that we are creating a parameterised type. Because the diamond isn't supported yet, if the length of targs doesn't equal the langt of tparams an error is thrown.