Defining fixpoints by well-founded recursion
A fixpoint F whose body is of the form { return E; } and
where either E is of the form x == 0 ? _ : _ or the fixpoint
has a decreases clause, is checked for contractiveness
(i.e. an application with measure M depends only on recursive
applications with measures 0 <= M' < M). If the check succeeds,
an unfolding lemma F_def is generated:

lemma void F_def(F_params)
    requires 0 <= measure(F_params);
    ensures F(F_params) == E;

Declaring the fixpoint as fixpoint_auto causes the lemma to be
generated as a lemma_auto. The resulting potential negative
SMT solver performance impact is the user's responsibility.