Print SHA-224 of release contents to log
The SHA-224 of a VeriFast named release binary package will be
different from that of the nightly release binary package printed
to the log, since the bin/VERSION file will be different. By printing
a separate SHA-224 of the release contents, the contents of a named
release can be checked against the build log.