stdio.h: add prototype for vfprintf
This exposed an incompleteness in stdarg.h, which is now fixed.

Also, this improves the definition of type `va_list` in stdarg.h.
It used to be defined as a pointer, but that is unsound.
(On my machine, sizeof(va_list) == 24.) It is now defined
abstractly. This required adding support for producing and
consuming points-to chunks for abstract types.