Check for null terminator in lemma string_to_chars
To the string_to_chars lemma in the post condition, added the assertion of the string having a null terminator at the end. This will satisfy the precondition of the null terminator check for chars_to_string. This addition helps when doing a string_to_chars to something and then later having a chars_to_string to convert it back.