Disallow reading from uninitialized fields
This breaking change (probably the biggest breaking change in the
history of VeriFast) is necessary to make VeriFast sound with respect to
the fact that C compilers perform optimizations that assume that the
program does not read uninitialized memory. (See e.g.
<https://www.ralfj.de/blog/2019/07/14/uninit.html> and
<https://godbolt.org/z/Y4x6cn6fq>.)

For each field f of type T of a struct S, VeriFast introduces a
predicate S_f(struct S *s, T v) that represents an initialized instance
of the field. This commit causes VeriFast to also introduce the
corresponding "underscore predicate" S_f_(struct S *s, option<T> v)
that represents a possibly-uninitialized instance of the field.

Writing consumes an underscore chunk and produces an initialized chunk;
reading requires an initialized chunk. Allocation produces an underscore
chunk and deallocation consumes an underscore chunk.

The syntax s->f |-> _ is now interpreted as denoting an underscore
chunk. The syntax s->f |-> ?_ can be used to denote an initialized
chunk.

See the discussion at #315.

TODO:
- Update the VeriFast Tutorial
- Make VeriFast sound with respect to pointer provenance and effective
  types
- Improve soundness regarding struct padding