Track pointer provenance
This breaking change is the first of a series of commits with the
goal of making VeriFast sound with respect to the fact that C
compilers perform optimisations that assume that pointers derived
from different allocations do not alias. See e.g.
<https://www.ralfj.de/blog/2018/07/24/pointers-and-bytes.html>
and <https://godbolt.org/z/nj6bhMq4o>.

From now on, a pointer is a pair consisting of a provenance (a
value of an abstract type `provenance`) and an address (a
uintptr_t). Pointer arithmetic does not affect the provenance.
Selecting a struct field affects the provenance, except if the
-assume_no_subobject_provenance flag is specified. Selecting a
union member currently does not affect the provenance. Casting
a pointer to uintptr_t simply returns its address; casting a
uintptr_t to a pointer yields a pointer with the same provenance
as the null pointer, which cannot be used to access any
memory (but you can work around this (unsoundly) by assuming that the
provenance of the allocation you are trying to access is also
the null pointer provenance). (If you have a pointer to some
allocation and you are casting a uintptr_t to a pointer to access
the same allocation, you can instead compute the difference
between the uintptr_t values and perform pointer arithmetic.
You can also use pointer arithmetic for pointer tagging as well;
see examples/mcas.)

See the discussion at #315.

TODO:
- Check pointer comparisons; these are only allowed if either they
  have the same provenance, or they point to the inside (not one-past)
  of live objects.
- If necessary, update the way object representation bytes are
  treated. (In CompCert and CH2O, object representation bytes are
  either poison, or plain integers, or pointer fragments.)
- (Eventually) support uintptr_t-to-pointer casts to construct
  pointers with non-null provenance.