Impose tight limits on pointer arithmetic
So far, the only check VeriFast did on pointer arithmetic is that the
address did not exceed the limits of uintptr_t.  With this breaking
change, VeriFast checks that the result of pointer arithmetic is within
the limits associated with the pointer's provenance. The limits of the
null pointer provenance coincide with those of uintptr_t; the limits of
other provenances are unspecified except that the minimum address is
greater than zero, the maximum address is at most UINTPTR_MAX, and if an
object is allocated at a given provenance, the addresses of that
object's representation bytes, as well as the one-past address, are
within the provenance's limits.

A very useful consequence is that the only pointer whose address is zero
is the null pointer.

To port your code, it should suffice to replace the incantation
0 <= p <= UINTPTR_MAX by pointer_within_limits(p) == true.