Check pointer limits when taking a field's address
Also checks them when taking the address of an array element. This
breaking change fixes a long-standing unsoundness.

This change should only cause proof breakage in cases where you
are taking the address of a field for which the thread does not
currently own a field chunk, *and* where the field is not the
first field of the struct.