Pointer equality tests possibly ignore provenance
A pointer equality test takes the pointers' provenances (i.e. the way
the pointers were derived) into account if the test is evaluated by the
compiler's alias analysis, and ignores them if it is evaluated at run
time. So if a test succeeds, all we know is that the addresses are
equal (not that the pointers are equal), and if it fails, all we know is
that the pointers are different (not that the addresses are different).

This may cause non-trivial proof breakage, so an -assume_no_provenance
flag is provided to allow this issue to be postponed.