Allow only the active union member to be accessed
In order to enable type-based alias analysis, accessing a member of a
union that is not the active member is considered undefined behavior
(with some exceptions). Initially, no member is active. This state is
represented in VeriFast as a char array whose length equals the size of
the union. A member can be activated by performing an activating
assignment to that member, i.e. an assignment whose left-hand-side is a
union member expression marked with /*@activating@*/. See
examples/unions.c. This consumes an initial part of the char array and
produces the member subobject at the subobject pointer (which has the
same address as the union object but a different provenance). Before the
union is deallocated, or another member is activated, the active member
must be deactivated using the new ghost command
'deactivate_union_member'. It consumes the specified subobject and
produces the corresponding char array.

If you do not care about being sound with respect to effective types
(i.e. if you compile with -fno-strict-aliasing), specify
the -assume_no_subobject_provenance command-line flag so that union
member pointers are equal to the union pointer itself.

TODO:
- Enforce effective types
- Allow reading of an inactive member through a union member expression.
  (This is called type punning.)
- Allow reading of a field F of an inactive union member M1 of type 'struct S1'
  if the active union member M2 is of type 'struct S2' and S1 and S2
  have a common initial sequence of fields and F is part of that common
  initial sequence.