Start enforcing effective types
This breaking change is a step towards making VeriFast sound with
respect to the fact that in the absence of the -fno-strict-aliasing
flag, compilers assume that an object of type T1 is accessed through an
lvalue of type T2 only if T1 == T2. This enables type-based alias
analysis.

Specifically, from now on, converting a chars(P, sizeof(int), _) chunk
to an integer(P, _) chunk is allowed only if has_type(P, &typeid(int))
== true.

TODO:
- Allow accesses through a pointer P of type intN_t * only if
  has_type(P, &typeid(intN_t)) == true. (integer_ chunks do not imply any
  particular has_type fact.)
- When P is of type struct S *, allow close_struct(P); only if
  has_type(P, &typeid(struct S)) == true.
- Currently, &typeid(T *) yields the same value of all types T. This is
  unsound with respect to the typeid operator's real behavior.