[Rust] T_own for tparams; open/close_fbc::<T>
This commit enables expressing ownership of values of type T where
T is a type parameter. This is accomplished by adding a T_own
parameter to each function with a type parameter T. When calling
such a function, the `own` predicate for the type argument is
passed in. Since struct own predicates take each field as a
separate argument, an S_own_ predicate that takes a struct value
is generated automatically for each struct S, expressed in terms
of the user-defined S_own predicate. S_own_ is passed when calling
functions with struct S as a type argument.

This commit also builds ghost commands

    open_full_borrow_content::<T>(t, l)

and

    close_full_borrow_content::<T>(t, l)

into the VeriFast engine. These consume/produce

    T_full_borrow_content(t, l)()

and produce/consume

    generic_points_to<T>(l, ?v) &*& T_own(l, v)

.