When consuming, distinguish typeid_env from env
Fixes an unsoundness in the consumption logic where the typeid for a
type parameter appearing in the type arguments of a heap chunk was
looked up in the environment used to evaluate the expressions appearing
in the chunk assertion. The type parameters in scope in the chunk
assertion should not be confused with the toplevel ones. Now looks these
up in the toplevel typeid environment.