[C++] `NodeListSerializer`: list serialization
This commit adds a generic `NodeListSerializer` class to support
serialization for lists of nodes for which the length is not known
in advance.

It does so by providing an API to iteratively serialize nodes.
Internally, it serializes the actual nodes to *orphans*. These orphans
can later adopted to an appropriate builder.