futex.h: disallow multiple futexes at same word
Fixes the unsoundness that results if multiple futexes, with
different values for `dequePost` or `callPermFunc`, are active
at the same time at the same address. In particular,
the spec of futex_wake_one is sound only if the thread being
awoken produces the same dequePost resource expected by the
awaking thread.