[Rust] _own takes struct value, not field values
For a struct S, the S_own predicate now takes a thread_id_t and an S value, instead of a thread_id_t and a value for each field of S.

This is in preparation for retiring _own predicates in favor of <T>.own. It also eliminates the difference between the treatment of structs whose body is known and those whose body is not known.

Also improves support for dereferencing struct pointers.