[Rust] Defaults: .own = false and .share = true
For each struct S defined in the crate being verified:
- if we do not find a user-provided definition of <S>.own, we define
  <S>.own(t, v) = false. If we do, and S is not trivially droppable
  and does not implement Drop, we generate the S_drop proof obligation.
- in all cases, we define S_full_borrow_content(t, l) and
  <S>.full_borrow_content = S_full_borrow_content
- if we do not find a user-provided definition of <S>.share, we define
  <S>.share(k, t, l) = true. If we do (and only then), we generate
  the share_full and share_mono proof obligations.

It follows that for each such struct S, definitions of predicates or predicate
constructors (as appropriate) S_own, S_full_borrow_content, and S_share
always exist with correct types. If the user does provide conflicting
definitions, this will cause a "Duplicate predicate/pred ctor" error
message.

It should now no longer be possible for generated contracts or the
generated S_full_borrow_content definition to be ill-typed. (These
were a big user experience problem because their source location
was just the struct and it was very hard for users to understand
what to do about them.) It should also no longer be possible for
the user to provide unsound definitions for these predicates, causing
the generated contracts to be unsound.

TODO: If S has public fields, additional proof obligations apply
which we do not yet check.