[Rust] Check for duplicate struct type pred defs
VeriFast now reports an error if a type predicate definition
for a struct type appears in a file that imports the struct
from another file, or if it contains two definitions of the
same type predicate for the same struct.

TODO: Check for duplicate type predicate definitions for
non-struct types.