[Rust] Always run in targeted mode
Since VeriFast uses rustc, which always assumes a specific target triple, a VeriFast run is sound only for that target triple. From now on, VeriFast mentions the target triple in the success message.

Also, a default C target (aka data model) is now assumed if none is specified: LLP64 on Windows and LP64 on Linux and MacOS. The data model is also mentioned in the success message.

This resolves some TODOs in the examples.

Fixes #579