[Rust] Support elided lifetimes
They get a unique name of the form '_N where N is their local def id.

Also: when generating a contract for a safe function, and a parameter of type &mut T has an elided lifetime, and the function's return type has no elided lifetimes, then the full borrow is opened, i.e. instead of requiring the full borrow, we require and ensure the points-to and the OWN of the pointee.