[Rust] Generate init_ref_S proof obligation
This proof obligation proves that memory referred to by &S references created by unverified typechecked clients is immutable for the lifetime of the shared reference (except for the parts that are under an UnsafeCell).

Proved it for linked_list.rs and generic_pair.rs.

TODO: Prove this proof obligations for the other examples