[Rust] Treat S_sync PO pre semantically
The MIR translator computes, for each struct S defined by the program,
which type parameters of S are necessarily Send or Sync when S is Sync.
For each of these, it used to add a separate precondition to the S_sync
proof obligation. The problem with that is that the set computed by the
MIR translator is a conservative approximation, which does not take into
account the definitions of structs other than S. For example, this
analysis does not see that if RawVec<T, A> is Sync, then A is
necessarily Sync as well, because A only appears in the definition of
RawVec as part of type RawVecInner<A>, and the analysis does not look at
the definition of RawVecInner when analyzing struct S.

From now on, the precondition for S_sync is simply is_Sync(typeid(S)).
Separately, an is_Sync_S_lemma autolemma is generated that expresses the
information resulting from the MIR translator's analysis.  However, if S
has no explicit Sync impls, another lemma, called is_Sync_S_def, is now
also generated, stating is_Sync(S) == is_Sync(T1) && ... && is_Sync(Tn),
where T1, ..., Tn are the types of the fields of S. This means VeriFast
can now tell that is_Sync(RawVec<T, A>) implies is_Sync(A).

TODO: Generate an is_Sync_S_def lemma also if S has Sync impls.